+++ src/main/java/me/desht/pneumaticcraft/common/entity/EntityRegistrator.java
@@ -12,32 +12,10 @@ import static me.desht.pneumaticcraft.common.util.PneumaticCraftUtils.RL;
 
 public class EntityRegistrator {
     public static void init() {
-        // Entities
-        // parms: entity class, mobname (for spawners), id, modclass, max player
-        // distance for update, update frequency, boolean keep server updated
-        // about velocities.
         EntityRegistry.registerModEntity(RL("vortex"), EntityVortex.class, "Vortex", 0, PneumaticCraftRepressurized.instance, 80, 1, true);
         EntityRegistry.registerModEntity(RL("drone"), EntityDrone.class, "Drone", 1, PneumaticCraftRepressurized.instance, 80, 1, true);
         EntityRegistry.registerModEntity(RL("logistic_drone"), EntityLogisticsDrone.class, "logisticDrone", 2, PneumaticCraftRepressurized.instance, 80, 1, true);
         EntityRegistry.registerModEntity(RL("harvesting_drone"), EntityHarvestingDrone.class, "harvestingDrone", 3, PneumaticCraftRepressurized.instance, 80, 1, true);
         EntityRegistry.registerModEntity(RL("ring"), EntityRing.class, Names.MOD_ID + ".ring", 100, PneumaticCraftRepressurized.instance, 80, 1, true);
-
-        // Entity Eggs:
-        // registerEntityEgg(EntityRook.class, 0xffffff, 0x000000);
     }
-
-//    private static int getUniqueEntityId() {
-//        int startEntityId = 0;
-//        do {
-//            startEntityId++;
-//        } while (EntityList.getClassFromID(startEntityId) != null);
-//
-//        return startEntityId;
-//    }
-
-//    public static void registerEntityEgg(Class<? extends Entity> entity, int primaryColor, int secondaryColor) {
-//        int id = getUniqueEntityId();
-//        EntityList.idToClassMapping.put(id, entity);
-//        EntityList.entityEggs.put(id, new EntityEggInfo(id, primaryColor, secondaryColor));
-//    }
 }